stub internal package
