+++ src/common/Experiments/Experiments-Live/Experiments.cs
@@ -51,7 +51,7 @@ namespace Experimentation
                 var featureFlag = allFeatureFlags[0].KeySegments[1];
                 FeatureFlagValue = allFeatureFlags[0].GetStringValue();
 
-                PowerToysTelemetry.Log.WriteEvent(new OobeVariantAssignmentEvent() { AssignmentContext = assignmentContext });
+                PowerToysTelemetry.Log.WriteEvent(new OobeVariantAssignmentEvent() { AssignmentContext = assignmentContext, ClientID = AssignmentUnit });
             }
             catch (Exception)
             {
@@ -61,33 +61,31 @@ namespace Experimentation
 
         public string? FeatureFlagValue { get; set; }
 
-        private static IVariantAssignmentRequest GetVariantAssignmentRequest()
+        private string? AssignmentUnit { get; set; }
+
+        private IVariantAssignmentRequest GetVariantAssignmentRequest()
         {
             string? clientID = string.Empty;
 
             string workingDirectory = Environment.CurrentDirectory;
+            var exeDir = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+            string settingsPath = @"AppData\Local\Microsoft\PowerToys\settings.json";
+            string jsonFilePath = Path.Combine(exeDir, settingsPath);
 
-            string? exeDir = Directory.GetParent(workingDirectory)?.Parent?.Parent?.FullName;
-            string settingsPath = @"src\settings-ui\Settings.UI.UnitTests\BackwardsCompatibility\TestFiles\v0.22.0\Microsoft\PowerToys\settings.json";
-            if (exeDir != null)
+            string json = File.ReadAllText(jsonFilePath);
+            var jsonDictionary = JsonSerializer.Deserialize<Dictionary<string, object>>(json);
+            if (jsonDictionary != null)
             {
-                string jsonFilePath = Path.Combine(exeDir, settingsPath);
-
-                string json = File.ReadAllText(jsonFilePath);
-                var jsonDictionary = JsonSerializer.Deserialize<Dictionary<string, object>>(json);
-                if (jsonDictionary != null)
+                if (!jsonDictionary.ContainsKey("clientid"))
                 {
-                    var jsonClientID = jsonDictionary["clientid"].ToString();
-
-                    if (jsonClientID == string.Empty)
-                    {
-                        jsonDictionary["clientid"] = Guid.NewGuid().ToString();
-                        string output = JsonSerializer.Serialize(jsonDictionary);
-                        File.WriteAllText(jsonFilePath, output);
-                    }
-
-                    clientID = jsonDictionary["clientid"]?.ToString();
+                    jsonDictionary.Add("clientid", string.Empty);
+                    jsonDictionary["clientid"] = Guid.NewGuid().ToString();
+                    string output = JsonSerializer.Serialize(jsonDictionary);
+                    File.WriteAllText(jsonFilePath, output);
                 }
+
+                clientID = jsonDictionary["clientid"]?.ToString();
+                AssignmentUnit = clientID;
             }
 
             return new VariantAssignmentRequest
